Split accounting reconciliation view model tests
